Seperated database and logger init
