+++ internal/server/server.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"log/slog"
 	"net/http"
@@ -37,6 +36,8 @@ type Opts func(s *Server)
 
 func NewServer(opts ...Opts) (*Server, error) {
 	s := defaultServer()
+
+	s.Logger = newLogger()
 
 	db, err := newDB(s.Logger)
 	if err != nil {
@@ -90,6 +91,21 @@ func ServerFromEnvs() (*Server, error) {
 	return NewServer(opts...)
 }
 
+func defaultServer() *Server {
+	return &Server{
+		Server: &http.Server{
+			Addr:         ":9001",
+			ReadTimeout:  time.Second,
+			WriteTimeout: 10 * time.Second,
+			IdleTimeout:  20 * time.Second,
+		},
+		Configs: &Configs{
+			Name: "todoApp",
+			Env:  "dev",
+		},
+	}
+}
+
 func loadEnvVars() []Opts {
 	var opts []Opts
 
@@ -108,70 +124,12 @@ func loadEnvVars() []Opts {
 		opts = append(opts, WithEnv(env))
 	}
 
-	readTimeout := getEnvAsInt("READ_TIMEOUT", 180)   // Default to 3 minutes
-	writeTimeout := getEnvAsInt("WRITE_TIMEOUT", 180) // Default to 3 minutes
-	idleTimeout := getEnvAsInt("IDLE_TIMEOUT", 300)   // Default to 5 minutes
+	readTimeout := getEnvAsInt("READ_TIMEOUT", 10)   // Default to 10 second
+	writeTimeout := getEnvAsInt("WRITE_TIMEOUT", 20) // Default to 20 second
+	idleTimeout := getEnvAsInt("IDLE_TIMEOUT", 30)   // Default to 30 second
 
 	opts = append(opts, WithTimeouts(readTimeout, writeTimeout, idleTimeout))
 	return opts
-}
-
-func defaultServer() *Server {
-	return &Server{
-		Server: &http.Server{
-			Addr:         ":9001",
-			ReadTimeout:  time.Second,
-			WriteTimeout: 10 * time.Second,
-			IdleTimeout:  20 * time.Second,
-		},
-
-		Configs: &Configs{
-			Name: "todoApp",
-			Env:  "dev",
-		},
-
-		Logger: newLogger(),
-	}
-}
-
-func newLogger() *slog.Logger {
-	logger := slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{AddSource: false}))
-
-	slog.SetDefault(logger)
-
-	return logger
-}
-
-func newDB(logger *slog.Logger) (*sqlitecloud.SQCloud, error) {
-	config := sqlitecloud.SQCloudConfig{
-		Host:     os.Getenv("DB_HOST"),
-		Port:     getEnvAsInt("DB_PORT", 8860),
-		Database: os.Getenv("DB_NAME"),
-		ApiKey:   os.Getenv("DB_APIKEY"),
-		MaxRows:  getEnvAsInt("DB_MAXROWS", 20),
-	}
-
-	isSecure, err := strconv.ParseBool(os.Getenv("DB_SECURE_FLAG"))
-	if err != nil {
-		return nil, err
-	}
-
-	config.Secure = isSecure
-
-	sqcl := sqlitecloud.New(config)
-
-	if err := sqcl.Connect(); err != nil {
-		logger.Error(err.Error())
-		return nil, err
-	}
-
-	if !sqcl.IsConnected() {
-		return nil, fmt.Errorf("not able to connect to database")
-	}
-
-	logger.Info("DB connected successfully")
-
-	return sqcl, nil
 }
 
 func getEnvAsInt(key string, defaultValue int) int {
